Moves methods and tests to Keys module
